requirements.txt
datetime
intersight
prettytable
python-dotenv
urllib3